Rename Iconstore folder
