+++ app/cdap/components/ConnectionList/IconStore/CGSIcon.tsx
