+++ app/cdap/components/ConnectionList/IconStore/WrangleIcon.tsx
